added a "makeEntryCopies" function to create copies of entries for front end
