+++ src/App.tsx
@@ -13,6 +13,7 @@ import { setEntries } from "./store/entriesSlice";
 import {addDays, addMonths} from 'date-fns';
 import { setReoccurEntries } from "./store/reoccurEntriesSlice";
 import { EntryAttributes } from './../server/db/models/Entry.model';
+import makeEntryCopies from "./../src/helpers/makeEntryCopies";
 
 
 export const ColorModeContext = React.createContext({
@@ -140,45 +141,14 @@ function App() {
   const reoccurEntriesFetch = useCallback(async() => {
     const token = window.localStorage.getItem("token");
     if(token){
-      const response = await axios.get("/api/entries", {
+      const entries = await axios.get("/api/entries", {
         headers: {
           authorization: `Bearer ${token}`,
         },
       });
-      let newEntries: EntryAttributes[] = [];
-      response.data[0].forEach((entry: EntryAttributes) => {
-        let newDate = new Date(entry.start);
-        if (entry.frequency === "Monthly") {
-          for (let i = 0; i <= 12; i++) {
-            let newEntry = structuredClone(entry);
-            newEntry.start = newDate.toISOString();
-            newEntries = [...newEntries,newEntry];
-            newDate = addMonths(newDate,1);
-          };
-        };
-        if (entry.frequency === "Bi-Weekly") {
-          for (let i = 0; i <= 26; i++) {
-            let newEntry = structuredClone(entry);
-            newEntry.start = newDate.toISOString();
-            newEntries = [...newEntries,newEntry];
-            newDate = addDays(newDate,14);
-          };
-        };
-        if (entry.frequency === "Weekly") {
-          for (let i = 0; i <= 52; i++) {
-            let newEntry = structuredClone(entry);
-            newEntry.start = newDate.toISOString();
-            newEntries = [...newEntries,newEntry];
-            newDate = addDays(newDate,7);
-          };
-        };
-        if (entry.frequency === "ByDate") {
-          let newEntry = structuredClone(entry);
-          newEntries = [...newEntries,newEntry];
-        };
-      });
-      dispatch(setReoccurEntries(newEntries));
-    }
+      const entryCopies = makeEntryCopies(entries.data[0]);
+      dispatch(setReoccurEntries(entryCopies));
+    };
   },[dispatch]);
 
   const theme = React.useMemo(() => createTheme(getDesignTokens(mode)), [mode]);
